+++ packages/desktop/lib/electron.ts
@@ -1,3 +1,3 @@
-import { IPlatform } from 'shared/lib/typings/platform'
+import { IPlatform } from '@lib/typings/platform'
 
 export const Electron: IPlatform = window['__ELECTRON__']
